fix: theme reestructurated to reduce build problems
